# Optimize JoinIDPath helper on metautil