@xiaofan-luan 2023-06-09
Hi @matiasinsaurralde
Sorry I missed the PR
Two things before we merge.
1) we need more test to verify these two fuctions are actually equals, although it looks good to me.
2) b.Grow(len(ids) * 19) where is the magic number 19 comes from?