@xiaofan-luan 2023-06-09
this is the length of strconv.FormatInt(id, 10)?