@matiasinsaurralde 2023-06-09
Hi @xiaofan-luan! I've added a detailed explanation here: https://go.dev/play/p/-zHvbZTHCef

To summarize:
- The iterator in `JoinIDPath` writes an int64 in string format plus a separator for all elements that are larger than 0.
- The maximum positive value of an int64 is `999999999999999999`, in stringified form this is 18 bytes.
- So if we want to use `Grow` we need to preallocate at least 18 bytes. We add one more byte for the separator that's added for all elements that are larger than 0, that's 19.